fix overrun

+++ resource/mesh.h
@@ -53,7 +53,7 @@ struct mesh {
   unsigned int m_vbo;
   unsigned int m_vao;
   uint32_t m_num_vertices;
-  size_t m_num_indices;
+  uint32_t m_num_indices;
   struct material m_material;
   struct vertex *mp_vertices;
   uint32_t *mp_indices;
